Update config to suppport eslint-plugin-react version 6.x
eslint-plugin-react had a major version update recently and it broke the setup
instructions for the plugin because the rules in this repository are no longer
compatible.

In particular:
* The allow-in-func setting for no-did-mount-set-state was removed, since it's
  now the default.
* The wrap-multilines rule was renamed to jsx-wrap-multilines. This resulted in
  a warning.

To test this change, I made a simple project with an .eslintrc that just
includes this plugin, and I followed the steps in the README. Without doing
anything special, it broke when I ran eslint, and after `npm link`ing this
change, it works.

Note that I think this breaks this plugin when used against eslint-plugin-react
versions prior to 6.0.0. I imagine this kind of problem could be avoided using
peer dependencies like eslint-config-airbnb does. In the meantime, this change
seems reasonable because probably people will be updating all lint-related
dependencies at once anyway.